+++ StartService/StartCommand.cs
@@ -1,9 +1,9 @@
 ﻿using Telegram.Bot.Types.ReplyMarkups;
-using TgBotPractice.DataBase;
 namespace TgBotPractice;
 using Telegram.Bot;
 using Telegram.Bot.Types;
 using Telegram.Bot.Types.Enums;
+using TgBotPractice.DataBase;
 
 public class StartCommand
 {
@@ -36,35 +36,10 @@ public class StartCommand
     💡 <i>Разработчик: @lvdshka
     Версия: 1.0.2</i>
     """;
-
-    public async Task DBCheck(Message msg)
-    {
-        using (ApplicationContext db = new ApplicationContext())
-        {
-            Human user = new Human { ChatId = msg.Chat.Id, City = String.Empty, Autosend = false, IsAdmin = false };
-            if (db.Users.Any(u => u.ChatId == user.ChatId))
-            {
-                Console.WriteLine($"ChatId: {user.ChatId} already exists.");
-            }
-            else
-            {
-                await db.Users.AddAsync(user);
-                await db.SaveChangesAsync();
-                Console.WriteLine($"ChatId: {user.ChatId} has been added.");
-            }
-
-            var users = db.Users.ToList();
-            foreach (var u in users)
-            {
-                Console.WriteLine(u.ChatId);
-            }
-
-        }
-    }
     
     public async Task StartCmd(ITelegramBotClient botClient, Message msg, UpdateType type)
     {
-        await DBCheck(msg);
+        await DbMethods.DBCheck(msg);
         var keyboard = new InlineKeyboardMarkup(new[]
         {
             new []
